+++ CHANGELOG.md
@@ -7,6 +7,8 @@ phpSysInfo 3.4.4
 ----------------
  - [ADD] Proxmox to detected distros
 
+ - [UPD] IPMIcfg sensor program - pminfo
+
 phpSysInfo 3.4.3
 ----------------
  - [ADD] Linux Lite, BigLinux, Linspire, Tuxedo, Armbian, Photon, RebeccaBlackOS, QuemOS, dahliaOS, openEuler, Runtu, risiOS, Nobara, Fatdog, CachyOS, Snal, Feren, Archman, Elive, RoboLinux, ExTiX, Kaisen, LXLE, Neptune, Regata, Archcraft, Kodachi, Bodhi, Siduction, Vanilla, Q4OS, Guix, XeroLinux, OpenWRT, Gnoppix, Uruk, Laclin, Clear, Sparky and Venom Linux to detected distros